requirements.txt
Flask
yfinance
requests
psycopg2-binary
